+++ src/ext/java/org/opentripplanner/ext/restapi/mapping/FareMapper.java
@@ -11,7 +11,6 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 import org.opentripplanner.ext.restapi.model.ApiCurrency;
-import org.opentripplanner.ext.restapi.model.ApiFareComponent;
 import org.opentripplanner.ext.restapi.model.ApiFareProduct;
 import org.opentripplanner.ext.restapi.model.ApiFareQualifier;
 import org.opentripplanner.ext.restapi.model.ApiItineraryFares;
@@ -24,7 +23,6 @@ import org.opentripplanner.model.fare.ItineraryFares;
 import org.opentripplanner.model.fare.RiderCategory;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.model.plan.Leg;
-import org.opentripplanner.routing.core.FareComponent;
 import org.opentripplanner.transit.model.basic.Money;
 
 public class FareMapper {
@@ -38,11 +36,10 @@ public class FareMapper {
   public ApiItineraryFares mapFare(Itinerary itinerary) {
     var fares = itinerary.getFares();
     Map<String, ApiMoney> apiFare = toApiMoneys(fares);
-    Map<String, List<ApiFareComponent>> apiComponent = toApiFareComponents(fares);
 
     return new ApiItineraryFares(
       apiFare,
-      apiComponent,
+      Map.of(),
       toApiFareProducts(fares.getItineraryProducts()),
       toApiLegProducts(itinerary, fares.getLegProducts())
     );
@@ -105,17 +102,6 @@ public class FareMapper {
     }
   }
 
-  private Map<String, List<ApiFareComponent>> toApiFareComponents(ItineraryFares fare) {
-    return fare
-      .getFareTypes()
-      .stream()
-      .map(key -> {
-        var money = fare.getComponents(key).stream().map(this::toApiFareComponent).toList();
-        return new SimpleEntry<>(key, money);
-      })
-      .collect(Collectors.toMap(e -> e.getKey().name(), Entry::getValue));
-  }
-
   private Map<String, ApiMoney> toApiMoneys(ItineraryFares fare) {
     return fare
       .getFareTypes()
@@ -138,9 +124,5 @@ public class FareMapper {
         c.getSymbol(locale)
       )
     );
-  }
-
-  private ApiFareComponent toApiFareComponent(FareComponent m) {
-    return new ApiFareComponent(m.fareId(), null, toApiMoney(m.price()), m.routes());
   }
 }