+++ src/main/java/org/opentripplanner/model/plan/ItinerarySortKey.java
@@ -2,13 +2,13 @@ package org.opentripplanner.model.plan;
 
 import java.time.Instant;
 import org.opentripplanner.framework.tostring.ValueObjectToStringBuilder;
-import org.opentripplanner.routing.algorithm.filterchain.comparator.SortOrderComparator;
-import org.opentripplanner.routing.algorithm.filterchain.filter.SortingFilter;
 
 /**
  * This interface is used to sort itineraries and other instances that we might want to sort among
- * itineraries. It is used in the {@link SortingFilter} as defined by the
- * {@link SortOrderComparator}.
+ * itineraries. It is used in the
+ * {@link org.opentripplanner.routing.algorithm.filterchain.framework.filter.SortingFilter}
+ * as defined by the
+ * {@link org.opentripplanner.routing.algorithm.filterchain.framework.sort.SortOrderComparator}.
  * <p>
  * The methods in this interface are NOT documented here, but in the Itinerary class. To keep it simple, this
  * interface should be kept in sync with method names in the itinerary.