+++ src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/request/RaptorRoutingRequestTransitDataCreator.java
@@ -147,7 +147,7 @@ class RaptorRoutingRequestTransitDataCreator {
             tripPattern.getAlightingPossible(),
             BoardAlight.ALIGHT
           ),
-          priorityGroupConfigurator.lookupTransitGroupPriorityId(tripPattern)
+          priorityGroupConfigurator.lookupTransitGroupPriorityId(tripPattern.getPattern())
         )
       );
     }