+++ src/test/java/org/opentripplanner/generate/doc/VehicleParkingDocTest.java
@@ -33,7 +33,7 @@ public class VehicleParkingDocTest {
   public void updateVehicleParkingDoc() {
     NodeAdapter node = readVehicleUpdaters();
 
-    // Read and close inout file (same as output file)
+    // Read and close input file (same as output file)
     String template = readFile(TEMPLATE);
     String original = readFile(OUT_FILE);